refactor : lottogame Divide into detailed class
