+++ src/main/java/lotto/game/controller/WalletController.java
@@ -1,0 +1,18 @@
+package lotto.game.controller;
+
+import lotto.domain.Wallet;
+import lotto.service.LottoProduceService;
+import lotto.vo.Money;
+
+public class WalletController {
+
+    private final LottoProduceService lottoProduceService;
+
+    public WalletController(LottoProduceService lottoProduceService) {
+        this.lottoProduceService = lottoProduceService;
+    }
+
+    public Wallet make(Money money) {
+        return lottoProduceService.produceLotto(money);
+    }
+}
